pkg: extract the tlog
Move the trivial logging (tlog) facilities out from the
deployer package into its own little package, to simplify
the dependency graph and to make it less awkward.

Trivial renames, but a lot of them.
