+++ pkg/commands/deploy.go
@@ -19,10 +19,10 @@ package commands
 import (
 	"log"
 
-	"github.com/fromanirh/deployer/pkg/deployer"
 	"github.com/fromanirh/deployer/pkg/deployer/api"
 	"github.com/fromanirh/deployer/pkg/deployer/rte"
 	"github.com/fromanirh/deployer/pkg/deployer/sched"
+	"github.com/fromanirh/deployer/pkg/tlog"
 
 	"github.com/spf13/cobra"
 )
@@ -54,7 +54,7 @@ func NewRemoveCommand(commonOpts *CommonOptions) *cobra.Command {
 		Use:   "remove",
 		Short: "remove the components and configurations needed for topology-aware-scheduling",
 		RunE: func(cmd *cobra.Command, args []string) error {
-			la := deployer.NewLogAdapter(commonOpts.Log, commonOpts.DebugLog)
+			la := tlog.NewLogAdapter(commonOpts.Log, commonOpts.DebugLog)
 			var err error
 			err = sched.Remove(la, sched.Options{
 				Platform:         commonOpts.Platform,
@@ -99,7 +99,7 @@ func NewDeployAPICommand(commonOpts *CommonOptions, opts *deployOptions) *cobra.
 		Use:   "api",
 		Short: "deploy the APIs needed for topology-aware-scheduling",
 		RunE: func(cmd *cobra.Command, args []string) error {
-			la := deployer.NewLogAdapter(commonOpts.Log, commonOpts.DebugLog)
+			la := tlog.NewLogAdapter(commonOpts.Log, commonOpts.DebugLog)
 			if err := api.Deploy(la, api.Options{Platform: commonOpts.Platform}); err != nil {
 				return err
 			}
@@ -115,7 +115,7 @@ func NewDeploySchedulerPluginCommand(commonOpts *CommonOptions, opts *deployOpti
 		Use:   "scheduler-plugin",
 		Short: "deploy the scheduler plugin needed for topology-aware-scheduling",
 		RunE: func(cmd *cobra.Command, args []string) error {
-			la := deployer.NewLogAdapter(commonOpts.Log, commonOpts.DebugLog)
+			la := tlog.NewLogAdapter(commonOpts.Log, commonOpts.DebugLog)
 			return sched.Deploy(la, sched.Options{
 				Platform:         commonOpts.Platform,
 				WaitCompletion:   opts.waitCompletion,
@@ -133,7 +133,7 @@ func NewDeployTopologyUpdaterCommand(commonOpts *CommonOptions, opts *deployOpti
 		Use:   "topology-updater",
 		Short: "deploy the topology updater needed for topology-aware-scheduling",
 		RunE: func(cmd *cobra.Command, args []string) error {
-			la := deployer.NewLogAdapter(commonOpts.Log, commonOpts.DebugLog)
+			la := tlog.NewLogAdapter(commonOpts.Log, commonOpts.DebugLog)
 			return rte.Deploy(la, rte.Options{
 				Platform:         commonOpts.Platform,
 				WaitCompletion:   opts.waitCompletion,
@@ -151,7 +151,7 @@ func NewRemoveAPICommand(commonOpts *CommonOptions, opts *deployOptions) *cobra.
 		Use:   "api",
 		Short: "remove the APIs needed for topology-aware-scheduling",
 		RunE: func(cmd *cobra.Command, args []string) error {
-			la := deployer.NewLogAdapter(commonOpts.Log, commonOpts.DebugLog)
+			la := tlog.NewLogAdapter(commonOpts.Log, commonOpts.DebugLog)
 			if err := api.Remove(la, api.Options{Platform: commonOpts.Platform}); err != nil {
 				return err
 			}
@@ -167,7 +167,7 @@ func NewRemoveSchedulerPluginCommand(commonOpts *CommonOptions, opts *deployOpti
 		Use:   "scheduler-plugin",
 		Short: "remove the scheduler plugin needed for topology-aware-scheduling",
 		RunE: func(cmd *cobra.Command, args []string) error {
-			la := deployer.NewLogAdapter(commonOpts.Log, commonOpts.DebugLog)
+			la := tlog.NewLogAdapter(commonOpts.Log, commonOpts.DebugLog)
 			return sched.Remove(la, sched.Options{
 				Platform:         commonOpts.Platform,
 				WaitCompletion:   opts.waitCompletion,
@@ -185,7 +185,7 @@ func NewRemoveTopologyUpdaterCommand(commonOpts *CommonOptions, opts *deployOpti
 		Use:   "topology-updater",
 		Short: "remove the topology updater needed for topology-aware-scheduling",
 		RunE: func(cmd *cobra.Command, args []string) error {
-			la := deployer.NewLogAdapter(commonOpts.Log, commonOpts.DebugLog)
+			la := tlog.NewLogAdapter(commonOpts.Log, commonOpts.DebugLog)
 			return rte.Remove(la, rte.Options{
 				Platform:         commonOpts.Platform,
 				WaitCompletion:   opts.waitCompletion,
@@ -199,7 +199,7 @@ func NewRemoveTopologyUpdaterCommand(commonOpts *CommonOptions, opts *deployOpti
 }
 
 func deployOnCluster(commonOpts *CommonOptions, opts *deployOptions) error {
-	la := deployer.NewLogAdapter(commonOpts.Log, commonOpts.DebugLog)
+	la := tlog.NewLogAdapter(commonOpts.Log, commonOpts.DebugLog)
 	if err := api.Deploy(la, api.Options{
 		Platform: commonOpts.Platform,
 	}); err != nil {
