+++ pkg/commands/render.go
@@ -23,11 +23,11 @@ import (
 	"github.com/spf13/cobra"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	"github.com/fromanirh/deployer/pkg/deployer"
 	"github.com/fromanirh/deployer/pkg/manifests"
 	"github.com/fromanirh/deployer/pkg/manifests/api"
 	"github.com/fromanirh/deployer/pkg/manifests/rte"
 	"github.com/fromanirh/deployer/pkg/manifests/sched"
+	"github.com/fromanirh/deployer/pkg/tlog"
 )
 
 type renderOptions struct{}
@@ -84,7 +84,7 @@ func NewRenderSchedulerPluginCommand(commonOpts *CommonOptions, opts *renderOpti
 				NodeResourcesNamespace: rteManifests.DaemonSet.Namespace,
 				PullIfNotPresent:       commonOpts.PullIfNotPresent,
 			}
-			la := deployer.NewLogAdapter(commonOpts.Log, commonOpts.DebugLog)
+			la := tlog.NewLogAdapter(commonOpts.Log, commonOpts.DebugLog)
 			return renderObjects(schedManifests.Update(la, updateOpts).ToObjects())
 		},
 		Args: cobra.NoArgs,
@@ -142,7 +142,7 @@ func renderManifests(cmd *cobra.Command, commonOpts *CommonOptions, opts *render
 		PullIfNotPresent:       commonOpts.PullIfNotPresent,
 	}
 
-	la := deployer.NewLogAdapter(commonOpts.Log, commonOpts.DebugLog)
+	la := tlog.NewLogAdapter(commonOpts.Log, commonOpts.DebugLog)
 	objs = append(objs, schedManifests.Update(la, schedUpdateOpts).ToObjects()...)
 
 	return renderObjects(objs)
