+++ pkg/deployer/rte/rte.go
@@ -20,6 +20,7 @@ import (
 	"github.com/fromanirh/deployer/pkg/deployer"
 	"github.com/fromanirh/deployer/pkg/deployer/platform"
 	rtemanifests "github.com/fromanirh/deployer/pkg/manifests/rte"
+	"github.com/fromanirh/deployer/pkg/tlog"
 )
 
 type Options struct {
@@ -29,7 +30,7 @@ type Options struct {
 	PullIfNotPresent bool
 }
 
-func Deploy(log deployer.Logger, opts Options) error {
+func Deploy(log tlog.Logger, opts Options) error {
 	var err error
 	log.Printf("deploying topology-aware-scheduling topology updater...")
 
@@ -64,7 +65,7 @@ func Deploy(log deployer.Logger, opts Options) error {
 	return nil
 }
 
-func Remove(log deployer.Logger, opts Options) error {
+func Remove(log tlog.Logger, opts Options) error {
 	var err error
 	log.Printf("removing topology-aware-scheduling topology updater...")
